+++ next.config.mjs
@@ -1,4 +1,6 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  experimental: { serverActions: { bodySizeLimit: "4mb" } },
+};
 
 export default nextConfig;